requirement.txt
Package           Version
----------------- -------------------
argcomplete       3.6.3
asteval           1.0.7
astropy           6.1.7
astropy-iers-data 0.2025.12.1.0.45.12
contourpy         1.3.2
cycler            0.12.1
dill              0.4.0
fonttools         4.61.0
kiwisolver        1.4.9
lmfit             1.3.4
matplotlib        3.10.7
numpy             2.2.6
packaging         25.0
pandas            2.3.3
pillow            12.0.0
pip               22.0.2
pyerfa            2.0.1.5
pyparsing         3.2.5
python-dateutil   2.9.0.post0
pytz              2025.2
PyYAML            6.0.3
scipy             1.15.3
setuptools        59.6.0
six               1.17.0
tomlkit           0.13.3
tzdata            2025.2
uncertainties     3.2.3
xmltodict         1.0.2
yq                3.4.3
